feat(projection): veil-tier hoist + gate composition — the two-gates hazard resolved (WO-41)
babylon.projection.veil relocates the doctrine Veil of Money verbatim
(monotonic set-membership gating, section 5d field->tier registry; web
shim is-identical, legacy test_veil.py green through it).
projection/fog/class_vision.py ports the EH Phase 2 desert/mud/water gate
as a pure non-mutating function (0.4-grid half-up quantization preserved;
the bridge wrapper delegates and its 12 behavioral tests stay green;
bridge-local constants removed as orphans).

projection/fog/precedence.py RESOLVES the documented fog+class-vision
collision hazard (fog/filter.py header, flagged unresolved since Track 1
Task 4): apply_political_gates pins vision-then-fog. Both gates are
restriction maps, so the composite never reveals more than either gate
alone — pinned by contract tests incl. reach-cannot-resurrect-withheld
and double-quantization determinism. 12 new tests; 221 across the
affected suites green; ledger +3 rows.
